Added Room DB to store predictions in

+++ data/src/main/java/com/jamieadkins/nfl/predictor/data/MatchesRepositoryImpl.kt
@@ -27,7 +27,7 @@ class MatchesRepositoryImpl @Inject constructor(
                        else -> MatchOutcomeEntity.Undecided
                    }
 
-                   MatchEntity(event.id.toLong(), homeTeam, awayTeam, outcome)
+                   MatchEntity(event.id, homeTeam, awayTeam, outcome)
                }
                listOf(GameWeekEntity(matches))
            }
